requirements.txt
langchain
langchain_community
langchain_groq
langchain_core
pypdf
sentence-transformers
streamlit
python-dotenv
langchain-huggingface
faiss-cpu